feat!: Remove fallback support to the "Deprecated" class
This class contained methods marked as deprecated since 2014 and used in only 3 places all over the code base.
The UPGRADING guide had instructions to migrate off them but some last callers remained.
These callers have since been patched.

This commit deletes both the actual "Deprecated" class and the fallback code which called into it.

Michael intended the Mime_Headers class to be lightweight and low level, having higher level API in separate classes.
The Deprecated class was also the only reason why Mime had a hard dependency on ListHeaders.
ListHeaders in turn had a much more reasonable dependency on Mime.
This circular dependency created some problems with composer.
